requirements.txt
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
pandas>=2.0
python-dotenv>=1.0
anthropic>=0.40        # Claude API (オプション: APIキーがある場合のみ使用)
